Fix!: Remove helm-shell.el in favour of helm-comint.el
For MELPA inclusion:

helm-shell.el with package-lint 20230905.629:

2 issues found:
31:0: error: Aliases should start with the package's prefix "helm-comint".
34:0: error: Aliases should start with the package's prefix "helm-comint".

I could rename the defaliases in helm-shell.el but as Pierre Neidhardt had made
clear; helm-shell.el is there for backwards compatability, as helm-comint.el is
to be used. To get this package onto MELPA, I'm making this breaking change.
I apologize in advance for any broken configs.